backend/pyproject.toml
[tool.poetry]
name = "crypto-news-agent-backend"
version = "0.1.0"
description = "Real-time Crypto News Agent Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.116.1"
uvicorn = "^0.35.0"
langchain-core = "0.3.69"
langchain-openai = "0.3.28"
langchain-huggingface = "0.3.0"
sentence-transformers = "^5.0.0"
qdrant-client = "1.14.3"
sse-starlette = "2.4.1"
websockets = "^15.0.1"
httpx = "^0.28.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"
feedparser = "^6.0.0"

newsapi-python = "^0.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
